requirements.txt
python-binance
TA-lib
numpy
Flask